Move admin_required to utils
It will now be used in two places, and makes sense to
have it in the admin utils.
